コードの不具合を洗い出して直す（器の版 2・検証一式）
査読で出た指摘を全部直し、直した後の再査読の指摘（確定 5 件・危険 12 件）も直した。
符号化の結果は動かしていない（15 件中 13 件がバイト数まで同じ、2 件は元のヘッダの
欄を戻すための +10 / +3 byte）。記録は results/defects.md。

- 並行性: 候補名の静的変数・tan 表の遅延初期化の競合、字母の使い回しの鍵、
  波の途中の f.col への書き込み
- 可逆性: LAS ヘッダの範囲・点数・GUID・ユーザーデータ・波形の位置、未記述の
  ExtraBytes、名前の衝突、空の ExtraBytes VLR、1.3 の波形データ
- 壊れた入力: 器・走査の流れの長さ検査、巨大な確保、計画の JSON、分母 0、PLY、KITTI
- 失敗の扱い: 作業用の名前に書いて検証後に移す、列の仕事の例外、main で例外を受ける
- 検証: pack が unpack --las と同じ道で書き戻して元と照合する、verify_suite.py、
  regress_fixes.py（作り物 42 項目）
- 未定義動作: -fwrapv、-ffp-contract=off、128 bit の線形検出、64 bit の数え上げ
- 非可逆: 誤差を利用者が受け取る float32 の値で測る（上限を 0.15% 超えていた）
- 器の版を 2 に上げた（封筒の末尾を札つきの拡張に改めたため）
- 文書の索引 results/INDEX.md と、履歴の文書に札を付けた

これまで未コミットだった 09-22〜23 の改良（符号つき文脈・パルス構造・曲面予測・
粗い格子・全次元保存など）もここに含む。

+++ cpp/include/pcc/scanmodel.hpp
@@ -43,7 +43,8 @@ inline int64_t med3(int64_t a, int64_t b, int64_t c) {
     return b;
 }
 struct MedPred {
-    int64_t prev = 0, d[3] = {0, 0, 0}; int k = 0;
+    // k は数え続けるので 64 bit にする（int では 2^31 点で負になり d[k % 3] が範囲外）
+    int64_t prev = 0, d[3] = {0, 0, 0}; uint64_t k = 0;
     inline int64_t predict() const {
         return prev + (k >= 3 ? med3(d[0], d[1], d[2]) : (k ? d[(k - 1) % 3] : 0));
     }
